Remove unused variable. Change type of local variable to avoid compiler warning.
git-svn-id: https://tinymux.googlecode.com/svn/branches/dev_brazil@1226 d1b986fa-651c-0410-a323-35a8662cf44d

+++ mux/src/bsd.cpp
@@ -3112,8 +3112,8 @@ static void process_input_helper(DESC *d, char *pBytes, int nBytes)
         d->raw_input_at = d->raw_input->cmd;
     }
 
-    int nInputBytes = 0;
-    int nLostBytes  = 0;
+    size_t nInputBytes = 0;
+    size_t nLostBytes  = 0;
 
     char *p    = d->raw_input_at;
     char *pend = d->raw_input->cmd + (LBUF_SIZE - sizeof(CBLKHDR) - 1);
